+++ environments/environment 2023-2-25 9-7-14.json
@@ -1,0 +1,1 @@
+{"roomNumber": 4, "floor": {"x": 116.83898065135406, "y": 103.46216148848009, "width": 110.06967676393774, "height": 159.6439956567735}, "R0": {"x": 33.38833971005678, "y": 231.26411475655254, "width": 115.29268332999834, "height": 119.44698930407327, "type": "bedroom", "children": [{"x": 62.64033620397835, "y": 337.170852548565, "width": 59.10203635506687, "height": 13.540251512060808, "type": "wardrobe", "orientation": "E", "children": []}, {"x": 43.9622581723051, "y": 231.26411475655254, "width": 50.080801082357, "height": 15.382080211219483, "type": "wardrobe", "orientation": "E", "children": []}], "door": {"x": 119.93785014224952, "y": 231.26411475655254, "width": 21.25, "height": 0}}, "R1": {"x": 68.82546194569433, "y": 48.32405805155843, "width": 86.26614704028621, "height": 86.98014582562271, "type": "bathroom", "children": [{"x": 145.96713517587193, "y": 120.45547881890207, "width": 8.898725058279089, "height": 14.848725058279088, "type": "toilet", "orientation": "S", "children": []}, {"x": 76.07815887141535, "y": 104.41620052754061, "width": 30.88800334964054, "height": 30.88800334964054, "type": "shower", "orientation": "S", "children": []}, {"x": 68.82546194569433, "y": 73.51466557726347, "width": 16.314125152604824, "height": 24.47118772890724, "type": "sink", "orientation": "E", "children": []}], "door": {"x": 121.62158041891806, "y": 135.30420387718115, "width": 21.25, "height": 0}}, "R2": {"x": 177.99310744398187, "y": 226.74530308528173, "width": 97.46028471519548, "height": 100.48956092034607, "type": "kitchen", "children": [{"x": 177.99310744398187, "y": 260.7698053947811, "width": 14.373196873481419, "height": 66.46505861084671, "type": "desk", "orientation": "S", "children": []}, {"x": 221.43474032436444, "y": 312.8616671321464, "width": 54.01865183481296, "height": 14.373196873481419, "type": "desk", "orientation": "E", "children": []}], "door": {"x": 196.2094002820846, "y": 226.74530308528173, "width": 21.25, "height": 0}}, "R3": {"x": 195.06661502659074, "y": 44.69284966029555, "width": 162.80452601812388, "height": 130.62638189944377, "type": "hall", "children": [{"x": 341.9654968508462, "y": 92.38150229440684, "width": 15.905644193868412, "height": 79.01322876019171, "type": "cupboard", "orientation": "S", "children": []}, {"x": 282.45465330507056, "y": 98.55423990827585, "width": 21.693808147667742, "height": 41.78178414530454, "type": "hallTable", "orientation": "", "children": [{"x": 289.2991882249799, "y": 140.3360240535804, "width": 8.004738307849056, "height": 8.004738307849056, "type": "chair", "orientation": "S"}, {"x": 304.14846145273833, "y": 115.44276282700359, "width": 8.004738307849056, "height": 8.004738307849056, "type": "chair", "orientation": "E"}, {"x": 289.2991882249799, "y": 90.5495016004268, "width": 8.004738307849056, "height": 8.004738307849056, "type": "chair", "orientation": "N"}, {"x": 274.4499149972215, "y": 115.44276282700359, "width": 8.004738307849056, "height": 8.004738307849056, "type": "chair", "orientation": "W"}]}], "door": {"x": 195.06661502659074, "y": 117.43686628761438, "width": 0, "height": 21.25}}}
