+++ environments/environment 2023-2-25 9-7-17.json
@@ -1,0 +1,1 @@
+{"roomNumber": 4, "floor": {"x": 61.99232762402637, "y": 127.85633470128388, "width": 202.44221277496263, "height": 179.3860475756395}, "R0": {"x": 222.9276534527865, "y": 67.23953905095031, "width": 126.29547802737635, "height": 113.02426627821067, "type": "bedroom", "children": [{"x": 222.9276534527865, "y": 67.8896244421609, "width": 14.043944936718493, "height": 57.91936636372068, "type": "wardrobe", "orientation": "S", "children": []}], "door": {"x": 222.9276534527865, "y": 149.082384060316, "width": 0, "height": 21.25}}, "R1": {"x": 160.50001875578528, "y": 241.2366491497012, "width": 89.18902460510495, "height": 80.30052702711966, "type": "bathroom", "children": [{"x": 160.50001875578528, "y": 251.72383062458005, "width": 31.768186961510068, "height": 31.768186961510068, "type": "shower", "orientation": "E", "children": []}], "door": {"x": 223.31772187046826, "y": 241.2366491497012, "width": 21.25, "height": 0}}, "R2": {"x": 4.614300108842701, "y": 265.7354953307209, "width": 98.88491446138619, "height": 92.58916809013309, "type": "kitchen", "children": [{"x": 88.54201782003183, "y": 284.6177211508999, "width": 14.957196750197058, "height": 73.70694226995408, "type": "desk", "orientation": "S", "children": []}, {"x": 46.75160286122212, "y": 343.36746667065694, "width": 41.79041495880971, "height": 14.957196750197058, "type": "desk", "orientation": "E", "children": []}, {"x": 4.614300108842701, "y": 300.1164911751049, "width": 14.957196750197058, "height": 58.208172245749076, "type": "desk", "orientation": "S", "children": []}], "door": {"x": 71.72940869689889, "y": 265.7354953307209, "width": 21.25, "height": 0}}, "R3": {"x": 15.418810944529742, "y": 10.618316245198843, "width": 158.52145561496724, "height": 158.74490540228754, "type": "hall", "children": [{"x": 52.75337221918581, "y": 53.73927653449357, "width": 40.64901390070622, "height": 24.13893815565489, "type": "hallTable", "orientation": "", "children": [{"x": 69.63524674327007, "y": 77.87821469014847, "width": 6.8852648525376985, "height": 6.8852648525376985, "type": "chair", "orientation": "S"}, {"x": 93.40238611989204, "y": 62.36611318605217, "width": 6.8852648525376985, "height": 6.8852648525376985, "type": "chair", "orientation": "E"}, {"x": 69.63524674327007, "y": 46.85401168195587, "width": 6.8852648525376985, "height": 6.8852648525376985, "type": "chair", "orientation": "N"}, {"x": 45.86810736664812, "y": 62.36611318605217, "width": 6.8852648525376985, "height": 6.8852648525376985, "type": "chair", "orientation": "W"}]}], "door": {"x": 93.86048160249709, "y": 169.3632216474864, "width": 21.25, "height": 0}}}
