+++ environments/environment 2023-2-25 9-7-18.json
@@ -1,0 +1,1 @@
+{"roomNumber": 4, "floor": {"x": 122.30586150897861, "y": 155.25733704016176, "width": 151.12768132242525, "height": 109.98660966605382}, "R0": {"x": 12.504434012114547, "y": 238.64968130228874, "width": 143.6359323929208, "height": 110.39572154080149, "type": "bedroom", "children": [{"x": 12.504434012114547, "y": 255.59426530557545, "width": 25.453125180707403, "height": 50.9531251807074, "type": "bed", "orientation": "N", "children": [{"x": 12.504434012114547, "y": 242.15836786416872, "width": 13.435897441406722, "height": 13.435897441406722, "type": "bedside", "orientation": ""}]}], "door": {"x": 156.14036640503537, "y": 241.2020690615518, "width": 0, "height": 21.25}}, "R1": {"x": 214.15030642649702, "y": 228.11727350775575, "width": 95.09764158604487, "height": 79.05626121875764, "type": "bathroom", "children": [], "door": {"x": 214.15030642649702, "y": 236.55713550730846, "width": 0, "height": 21.25}}, "R2": {"x": 246.83927742747701, "y": 61.068396728021064, "width": 96.25487250673997, "height": 120.78320571606753, "type": "kitchen", "children": [{"x": 273.9243843113569, "y": 61.068396728021064, "width": 69.16976562286013, "height": 14.742466350819505, "type": "desk", "orientation": "E", "children": []}, {"x": 328.3516835833975, "y": 75.81086307884057, "width": 14.742466350819505, "height": 79.75025186273382, "type": "desk", "orientation": "S", "children": []}, {"x": 263.60885090252725, "y": 98.81836410287387, "width": 19.389522774381774, "height": 26.886081805678167, "type": "table", "orientation": "", "children": [{"x": 269.2504807643646, "y": 125.70444590855203, "width": 8.106263050707037, "height": 8.106263050707037, "type": "chair", "orientation": "S"}, {"x": 282.998373676909, "y": 108.20827348035944, "width": 8.106263050707037, "height": 8.106263050707037, "type": "chair", "orientation": "E"}, {"x": 269.2504807643646, "y": 90.71210105216683, "width": 8.106263050707037, "height": 8.106263050707037, "type": "chair", "orientation": "N"}, {"x": 255.5025878518202, "y": 108.20827348035944, "width": 8.106263050707037, "height": 8.106263050707037, "type": "chair", "orientation": "W"}]}], "door": {"x": 246.83927742747701, "y": 160.1571677344366, "width": 0, "height": 21.25}}, "R3": {"x": 18.17455512115359, "y": 46.05886641816202, "width": 130.72557179175186, "height": 168.76129852325255, "type": "hall", "children": [{"x": 18.26909379758556, "y": 77.04343684862934, "width": 21.5913237756568, "height": 62.56422488134204, "type": "sofa", "orientation": "E", "children": []}], "door": {"x": 148.90012691290545, "y": 156.66733672778037, "width": 0, "height": 21.25}}}
