+++ environments/environment 2023-2-25 9-7-19.json
@@ -1,0 +1,1 @@
+{"roomNumber": 4, "floor": {"x": 113.1752359773018, "y": 142.95720323009442, "width": 191.44691746217543, "height": 160.99376336480015}, "R0": {"x": 31.391383755176065, "y": 268.27349063481097, "width": 117.46132818220933, "height": 102.56153408286441, "type": "bedroom", "children": [{"x": 70.31574046533424, "y": 268.27349063481097, "width": 21.470373023450122, "height": 46.97037302345012, "type": "bed", "orientation": "S", "children": [{"x": 54.54563823976474, "y": 268.27349063481097, "width": 15.770102225569499, "height": 15.770102225569499, "type": "bedside", "orientation": ""}]}], "door": {"x": 148.85271193738538, "y": 279.5277531747176, "width": 0, "height": 21.25}}, "R1": {"x": 212.32363381730354, "y": 268.2599947388201, "width": 76.47354243296448, "height": 75.75457300204808, "type": "bathroom", "children": [{"x": 228.52032462993, "y": 326.6705045112178, "width": 26.0160948444756, "height": 17.3440632296504, "type": "sink", "orientation": "N", "children": []}], "door": {"x": 245.8056410089426, "y": 268.2599947388201, "width": 21.25, "height": 0}}, "R2": {"x": 268.94467747939365, "y": 85.53188447677057, "width": 116.80156063774808, "height": 93.10279471340742, "type": "kitchen", "children": [{"x": 339.91966217360937, "y": 102.76097150298305, "width": 26.366075707139007, "height": 20.509896419717744, "type": "table", "orientation": "", "children": [{"x": 349.0251966849879, "y": 123.2708679227008, "width": 8.15500668438193, "height": 8.15500668438193, "type": "chair", "orientation": "S"}, {"x": 366.2857378807484, "y": 108.93841637065096, "width": 8.15500668438193, "height": 8.15500668438193, "type": "chair", "orientation": "E"}, {"x": 349.0251966849879, "y": 94.60596481860112, "width": 8.15500668438193, "height": 8.15500668438193, "type": "chair", "orientation": "N"}, {"x": 331.76465548922744, "y": 108.93841637065096, "width": 8.15500668438193, "height": 8.15500668438193, "type": "chair", "orientation": "W"}]}], "door": {"x": 268.94467747939365, "y": 150.13440277374124, "width": 0, "height": 21.25}}, "R3": {"x": 95.31515900461018, "y": 30.191741263157837, "width": 140.85139763062747, "height": 156.1263261559091, "type": "hall", "children": [{"x": 95.31515900461018, "y": 68.91201514799158, "width": 13.266488693123657, "height": 83.94244203181853, "type": "cupboard", "orientation": "S", "children": []}, {"x": 188.94007197615372, "y": 30.191741263157837, "width": 26.43234632930831, "height": 13.706490360899338, "type": "cupboard", "orientation": "E", "children": []}, {"x": 191.63893477220404, "y": 115.49830906792045, "width": 22.622197755172266, "height": 57.24252693828247, "type": "sofa", "orientation": "W", "children": []}], "door": {"x": 170.98845711749556, "y": 186.31806741906695, "width": 21.25, "height": 0}}}
